# Welcome Welcome to The Codebase.
One codebase is a singular codebase that serves as both a template to start a project as well as a tutorial on how to code. It does this by asserting that every set of changes is a cohesive lesson, and every lesson is taught as a series of changes. As a result, each branch of this codebase of this codebase can show you exactly how it got to the way it was by following each predecessor lesson.

## Documents
A repository is a series of documents. These documents should have structure and purpose.
  - [development](./development.md) - document responsible for getting a developer started in this repository
  - [thesaurus](./thesaurus.md) will provide definitions of important words