# Welcome Welcome to The Codebase.
One codebase is a singular codebase that serves as both a template to start a project as well as a tutorial on how to code. It does this by asserting that every set of changes is a cohesive lesson, and every lesson is taught as a series of changes. As a result, each branch of this codebase of this codebase can show you exactly how it got to the way it was by following each predecessor lesson.

## Course Overview
### Code
This is a course about code. Code represents formalized ideas. We can use computer code to tell a computer what to do. We can use legal code to tell a court how to rule. We use code to create  abstract structures that lead consistent results.

#### Fundamental Unit of Code
Code has some fundamental unit. That is, there is a point where we cannot breakdown code anymore. We can break up legal documents into sections that each serve a cohesive purpose. We can even go further. Sections can be broken up into paragraphs which may be further broken up into sentences and then clauses.

Can sentences/clauses be broken up into words and maintain a cohesive purpose? No. A single word cannot convey an idea. In the same way, the smallest unit of code will be enough to complete a purpose.

#### Software Engineering
Software Engineering is the process of breaking down a problem into smaller problems, solving those smaller problems, and then combining those smaller solutions into bigger problems. If the smaller problems cannot be solved, we can reapply our thought process of breaking them up.

### Philosophy
Every lesson directly improves a students ability to code.
 - We teach through Github as this tool is directly related to software development.
 - Links to articles are not given directly. Instead, we present a search result that will lead to the article.

#### Inspiration
This course was inspired by an article

### Takeaways
Upon completing this course, you will:

### Resources
#### Discord
#### Glossary
Two ways to navigate to it.

## Connect 4
As an exercise

## Next Lesson