# Welcome Welcome to The Coding Course.
The Coding Course is dedicated to explaining how to code. Code is a representation of formalized ideas. Just like legal code tells a court how to rule, computer code tells a computer how to execute tasks. These

## Codebase template
This course can also
It is also a codebase that can be used to start a project a singular codebase that serves as both a template to start a project as well as a tutorial on how to code.

In this course, does this by asserting that every set of changes is a cohesive lesson, and every lesson is taught as a series of changes. As a result, each branch of this codebase of this codebase can show you exactly how it got to the way it was by following each predecessor lesson.

## Course Overview
### Code

#### Fundamental Unit of Code
Code has some fundamental unit. That is, there is a point where we cannot breakdown code anymore. We can break up legal documents into sections that each serve a cohesive purpose. We can even go further. Sections can be broken up into paragraphs which may be further broken up into sentences and then clauses.

Can sentences/clauses be broken up into words and maintain a cohesive purpose? No. A single word cannot convey an idea. In the same way, the smallest unit of code will be enough to complete a purpose.

#### Software Engineering
Software Engineering is the process of breaking down a problem into smaller problems, solving those smaller problems, and then combining those smaller solutions into bigger problems. If the smaller problems cannot be solved, we can reapply our thought process of breaking them up.

### Philosophy
Every lesson directly improves a students ability to code.
 - We teach through Github as this tool is directly related to software development.
 - Links to articles are not given directly. Instead, we present a search result that will lead to the article.

#### Inspiration
This course was inspired by an article about a [statistic course](https://www.google.com/search?q=statistics+course+fractal) that was taught in a fractal building.

### Takeaways
Upon completing this course, you will:
 - practice researching technical topic and completing technical tutorials.

### Resources
#### Discord
#### Glossary
TODO add glossary.

## Connect 4
As an exercise, take the game Connect 4 and describe it. Would your description tell a new player all the rules.

## Next Lesson
Currently you are at lesson 0. To get to lesson 1, find the dropdown with `lesson-0` and change it to `lesson-1`

### Philosophy
Every lesson directly improves a students ability to code.
 - We teach through Github as this tool is directly related to software development.
 - Links to articles are not given directly. Instead, we present a search result that will lead to the article.

#### Inspiration
This course was inspired by an article about [statistics course](https://www.google.com/search?q=statistics+course+fractal) taught in a fractal pattern.

### Takeaways
Upon completing this course, you will:
 - search for documents
 - breakdown problems into smaller problems and then implement solutions via software
 - organize technical and non-technical solutions into coherent folder structures

### Resources
#### Discord
Join us on [discord](https://discord.gg/AJ6bswH5eg) on Wednesdays @7PM EST.

## Connect 4
As an exercise try to describe Connect 4 to someone who wants to play it but never has.

Connect4 is a two player game where players take turns dropping their colored pieces to the bottom of an incomplete column 7 column x 6 rows board game.
A player wins when they connect 4 pieces in one of three ways: vertically, horizontally, diagonally.

As a to-do list, this might look like:
 - two players -> tracker for turn (binary)
 - 7 column x 6 rows board game
   - dropping pieces to the bottom of an incomplete column
 - connect 4
   - vertically
   - horizontally
   - diagonally

## Next Lesson
Currently, you are on lesson 0. To go to the next lesson, lesson 1, find the button with `lesson-0`  and change it to lesson 1.