# List names Use a list to display the names. This list is a javascript data structure for holding multiple pieces of information. Here that information is the individual name.
The key `[` is used to denote the start of a list, `]` is intended to denote the end. In between is a comma separated list elements. The last element should also contain a comma to improve the symmetry of elements in a list. This avoids having to add a comma to a line not affected by the changes we are making.

When displaying the value. We get the value with `names[i]` where i is the index (position in the array).

To display javascript inside of the html, we need to used the escape keys: `{` denotes opening and `}` and to close. Any valid javascript can be escaped out and whatever the result of its evaluation gets displayed.

## Resources
 - [developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array]