# Welcome to Java Java seed code.
## .github/workflows
[Github Actions](https://www.google.com/search?q=github+actions) allow us to run code and get feedback on it. We can find multiple suggestions for [java github actions](https://www.google.com/search?q=github+actions). We picked one that implemented a [java github action with Maven](https://docs.github.com/en/actions/automating-builds-and-tests/building-and-testing-java-with-maven).

## .gitignore
We are version controlling the code with git. `.class` files are the output of java compilation. We do not want to keep output files, we only want to keep info on hot to create them which happens through a Dockerfile.

## Dockerfile
A Dockerfile

## main.java

Main executable. This is the entry point to the app when we are running it in it's final versions.

## Greet

## Dockerfile
Shamelessly stolen from https://hub.docker.com/_/openjdk