# Column Drop We want to take our Slots and turn them into a single column. When a player clicks on the Column, the slot furthest down is filled with a token from the current `playerTurn` and the `playerTurn` changes.
## TODOs
 - [ ] Make 6 slots
 - [ ] Align these 6 slots vertically
 - [ ] If a user Clicks a Column, a token with the current Players Turn will fill the bottom most slot
   - [ ] We need to capture a state for the tokens that are in the column
   - [ ] When we click on a Column we want to take the current `playerTurn` and add it to our stored toked
   - [ ] We want to ensure there is room in the column. When there is already 6 tokens, we don't want to do anything.
   - [ ] We want to remember to change the player turn
   - [ ] Instead of handling this event on the slot token, we should be handling it on the column itself. This will help us keep track of what to manipulate when we have many Columns

## Details
Given one slot, we want to make a column of them. To do this, we repeat the same html element that was used to create one. In order to ensure that our new element have a blue background, we want them to be nested inside a parent element with the `blue` `background-color` property.

In order for the elements to be displayed vertically, we want to change the `display` from it's default and to `inline-block`. `display` is an often used css property in arranging elements. If you are looking to arrange element it is worth trying different `display` values.

To store the values of tokens in the column, we need to create a state variable with a data structure that can store many values. We call `useState` with an initially empty Array defined by `[]`. We get back an array of two values: a variable with the value and a function for updating the value. We call these `tokens` and `changeTokens`.

We then want to use these to create another function called `addToken`. When this function is called we want to add the current `playerTurn` to the chips and update them. In order to do so, we create a new Array with all the values in token `...tokens` along with our desired `playerTurn`.

`addToken` should be an event handler on our Column.

We should use the values in `tokens` to determine what color should be displayed in a slot. The first value passed in at `tokens[0]` should correspond to what we are displaying in the bottom slot of the column. `tokens[1]` should correspond to our second from the bottom slot and so on.

## Further Reading
 - [https://www.google.com/search?q=js+data+structure+store+many+values](js data structure store many values) will lead to choosing Arrays. It is also worth thinking about why Arrays come up frequently: we often want to store many values.
 - [https://www.google.com/search?q=react+use+state](react use state) for reactjs documentation on how to handle state in react
 - [https://www.google.com/search?q=mozilla+display](mozilla display) for mozilla documentation on the display css tag